Move layout components to separate directory
