+++ src/components/layout/Footer.module.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use '../../variables' as *;
 
 footer {
   background-color: darken($color-light, 10%);
