+++ src/components/layout/Header.module.scss
@@ -1,4 +1,4 @@
-@use '../variables' as *;
+@use '../../variables' as *;
 
 nav {
   display: flex;
