+++ src/pages/Quote.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import PageWrapper from '../components/PageWrapper';
+import PageWrapper from '../components/layout/PageWrapper';
 
 class QuotePage extends React.Component {
   render() {
